[apollo-mock-client] Rename and support query, mutation, and subscription operations
