+++ packages/apollo-mock-client/README.md
@@ -1,0 +1,3 @@
+# ApolloMockClient
+
+An Apollo Client that allows mocking of payloads in response to operations, rather providing them all upfront. It is API-wise a port of [Relay’s RelayMocEnvironment](https://relay.dev/docs/guides/testing-relay-components/#relaymockenvironment-api-overview).
