+++ packages/apollo-mock-client/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@graphitation/apollo-mock-provider",
+  "name": "@graphitation/apollo-mock-client",
   "version": "0.1.0",
   "main": "./src/index.ts",
   "scripts": {
@@ -11,10 +11,15 @@
   },
   "devDependencies": {
     "@apollo/client": "^3.3.15",
+    "@graphitation/graphql-js-tag": "^0.1.0",
+    "@graphitation/graphql-js-operation-payload-generator": "^0.1.0",
     "@types/jest": "^26.0.22",
     "@types/react": "^17.0.3",
+    "@types/react-test-renderer": "^17.0.1",
     "graphql": "^15.5.0",
-    "monorepo-scripts": "*"
+    "monorepo-scripts": "*",
+    "react": "^17.0.2",
+    "react-test-renderer": "^17.0.2"
   },
   "publishConfig": {
     "main": "./lib/index.js",
