--- packages/apollo-mock-provider/README.md
@@ -1,3 +1,0 @@
-# ApolloMockProvider
-
-An Apollo Client provider that allows mocking of payloads in response to operations, rather providing them all upfront. It is API-wise a port of [Relay’s RelayMocEnvironment](https://relay.dev/docs/guides/testing-relay-components/#relaymockenvironment-api-overview).
