--- packages/apollo-mock-provider/src/__tests__/index.test.ts
@@ -1,7 +1,0 @@
-import { bar } from "..";
-
-describe("bar", () => {
-  it("returns a string", () => {
-    expect(typeof bar("test")).toBe("string");
-  });
-});
